backend/requirements.txt
aiohttp==3.7.3
Django==3.1.4
requests==2.25.0
aiofiles==0.6.0
genanki==0.9.1
djangorestframework==3.12.2
